// types/app.d.ts
export type CardType =
  | 'Loading'
  | 'Editor'
  | 'Diff'
  | 'Explorer'
  | 'Browser'
  | 'Branches'
  | 'Merge'
  | 'SourceControl';